[balsa] Fix 304WithBody test and add new one.
Add //test/common/http/http1:codec_impl_test test case for 304 response
with body for both BalsaParser and http-parser.  In this test, the body
is dispatched to the parser in the same Parser::execute() call as the
headers.

Modify ProtocolIntegrationTest.304WithBody by appending '\r\n' to the
end of the "body" of the 304 response, necessary for BalsaParser to
signal an error.

Signed-off-by: Bence Béky <bnc@google.com>