Fix PARTIAL_BUFFERED to support header modification
Modify PARTIAL_BUFFERED code path so that it does not resume the plugin
flow until either the body has been completely received OR if the body
limit has been exceeded.

This way, a processor using this mode can change headers or send an
immediate response after examining the body, unless the body is larger
than the limit.

Signed-off-by: Greg Brail <gbrail@users.noreply.github.com>