Call Parser::onMessageBegin() early.
Call Parser::onMessageBegin() as soon as the first byte of the message
is received, just like http-parser does.  Note that this does not
exactly match http-parser behavior, which does not call
Parser::onMessageBegin() if the first character cannot begin a valid
message (for example, request starts with a character that is not the
first character of any defined HTTP method).

Signed-off-by: Bence Béky <bnc@google.com>