Reduce VirtualHost memory utilization by avoiding CatchAllVirtualCluster instance when not needed.
This is a step towards more memory efficient config data structures. Issue #24154.

Currently an instance of CatchAllVirtualCluster is always created for
each vhost, but is only used when there are other virtual clusters (and
none of them matches). If there are no virtual clusters to begin with,
CatchAllVirtualCluster is a pure overhead.

This PR switches an instance to a unique_ptr and only creates one when
there are virtual clusters present.

Signed-off-by: Yury Kats <ykats@google.com>