ratelimit: allow metadata descriptors to be skipped
The default behaviour was to skip calling the rate limiting service
whenever the metadata key was not found and the default value was not
set. This was not correctly documented (nor tested) since it mentioned
that only the descriptor was skipped whereas the rate limiting service
was skipped altogether.

This adds a skip_if_absent field in the same spirit as for the request
headers action: if set to true it skips the descriptor but still calls
the rate limiting service, otherwise it skips the rate limiting service.

Note that the deprecated dynamic metadata action does not support this
field and defaults to false.

Signed-off-by: Thomas van Noort <thomas.vannoort@datadoghq.com>