Reduce Route/VirtualHost memory utilization by avoiding optional RateLimitPolicy instances when not needed.
Currently an instance of RateLimitPolicyImpl is created in each vhost
and route, even if no rate limiting is configured. This PR switches
instances to unique_ptr and only creates RateLimitPolicyImpl objects
when actually configured. Otherwise a singletong default policy is used.

This is a step towards more memory efficient config data structures. Issue #24154.

Similar change to Retry, Hedge and InternalRedirect policies was made in
PR #24182.

Signed-off-by: Yury Kats <ykats@google.com>