Reduce Route memory utilization by sharing the most common prefix path matchers
An analysis of a massively multi-tenant Envoy deployment shows that 75%
of path matchers have either "" or "/" as a prefix. This PR creates a shared instance for both, so that all the routes can re-use the same trivial matcher instead of creating new identical ones.

This is a step towards more memory efficient config data structures. Issue #24154.

Signed-off-by: Yury Kats <ykats@google.com>