flip RUNTIME_GUARD 'envoy.reloadable_features.thrift_connection_draining' to true (#24106)
Pinterest introduced this runtime_guard before. We've done tests in house and they were looking good, so we feel it is safe now to flip it to default true. We plan to remove this runtime_guard and make the feature default in half a year. 
Risk Level: low
Testing: done in house
Docs Changes: none
Release Notes: none
Platform Specific Features:  none

Signed-off-by: xianggao <xianggao@pinterest.com>