stats: Scope should not be a parent-class of Store (#23851)
Commit Message: Refactors so that Stats::Store is no longer a subclass of Stats::Scope. Previously it was, so that the root-scope was effectively the same thing as the Store. This causes issues and confusion because Stores are typically created directly as test-class member variables, but Scopes were mostly created as shared_ptr for various reasons. This refactor attempts to cleanly, but minimally, enable the memory management of Scopes to be strictly shared_ptr. It also separate two different concepts.

Additional Description:

This leaves behind a bit of technical debt to keep this refactor reviewable. This is tracked in #24007 and can be addressed in a follow-up PR that will be much larger than this one, but will be strictly a series of simple replacements (makeStats(store_) --> makeStats(*store_->rootScope()))

To reviewers: I recommend looking at scope.h and store.h carefully to see the real change that was done here. Then also look at isolated_store_impl.*, stat_test_utility.*, and the changes to the mocks. Finally look at test/integration/server.h for the change in the thread-safe wrapper around isolated store there. Everything else should be pretty quick to review once you have that context.

Risk Level: medium -- this is a pretty large refactor
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
Fixes: #23862

Signed-off-by: Joshua Marantz <jmarantz@google.com>