mobile: add `./remote` helper
Envoy Mobile project developers can use EngFlow's RBE to speed up and
offload local builds to EngFlow's remote execution cluster.

Most bazel commands should work by replacing `./bazelw` with `./remote`.

For example:

```console
$ ./remote build --config ios ios_dist
$ ./remote build --config android android_dist
```

With certificates places in the `mobile/tmp/certs` directory (which is
ignored by git), it should "just work".

Note that `protoc` commands fail with a "Bad CPU type in executable"
error message when using RBE on an M1 Mac with an arm64 bazel binary, so
to work around this we tell `bazelw` to use the amd64 version, which
appears to work just fine.

This depends on https://github.com/envoyproxy/envoy/pull/24501 if you're
on an arm64 host because all the remote executors are on amd64.

Signed-off-by: JP Simard <jp@jpsim.com>