mobile: Test counters use Stats::Store instead of admin interface
This commit changes the Envoy Mobile integration tests to use the
Stats::Store for retrieving counter values instead of making requests to
the admin interface. Now, stats are obtained from the stats store, not
from the admin interface, so it unblocks efforts around compiling out
the admin interface when building Envoy Mobile.

Signed-off-by: Ali Beyad <abeyad@google.com>