Reduce Route memory utilization by avoiding extra copies of matcher strings (#24614)
Multiple sub-classes of RouteEntryImplBase (such as
PrefixRouteEntryImpl) make two copies of the same match criterion string
-- one copy is kept inside the PathMatcher they create, and another is
kept as a class member. This PR avoids duplication by eliminating
the criteria as the class member, and using a reference to PathMather's copy.

This is a step towards more memory efficient config data structures. Issue #24154.

Signed-off-by: Yury Kats <ykats@google.com>