Reduce Route memory utilization by avoiding ConnectConfig instances when not needed (#24589)
Currently an instance of ConnectConfig (wrapped in optional) is created in each route, even if not configured. This PR switches the instance to unique_ptr and only creates ConnectConfig objects when actually configured.

To facilitate the refactoring, connectConfig() API changes from absl::optional<ConnectConfig>& to ConnectConfigOptRef. Most of the PR changes are to reflect that, updating tests and mocks to the new signature.

This is a step towards more memory efficient config data structures. Issue #24154.

Risk Level: low
Testing: updated
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a
Signed-off-by: Yury Kats <ykats@google.com>