Reduce Route memory utilization by avoiding `absl::optional` overhead.
`std::duration` values wrapped in an `absl::optional` take twice as
much memory as the durations themseves -- 16 bytes for each optional
milliseconds value. This PR removes the optional overhead by using
a sentinel -1 value to represent lack of the duration.

There are 6 optional durations on each Route, and 2 more in RetryPolicy,
so this is saving 64 bytes per route.

This is a step towards more memory efficient config data structures. Issue #24154.

Signed-off-by: Yury Kats <ykats@google.com>