Reduce Route memory utilization by avoiding redirect config overhead when no redirect is configured
Currently each Route instance incurs contains 10 member variables that
are representing `route.redirect()` config, whether the route has a
redirect or not. For routes without a redirect, that's 6 empty strings,
1 pointer and 3 booleans.

This PR packs all redirect-related config into a separate data structure
and only creates an instance of it when the route has a redirect. This
is saving 150 bytes for any route that does not have a redirect.

This is a step towards more memory efficient config data structures. Issue #24154.

Signed-off-by: Yury Kats <ykats@google.com>