Copy stream info for deferred logging instead of using shared ptr
Here we revert the change of wrapping the active stream's stream info
into a shared pointer. Instead, we copy the stream info to explicitly
ensure that all members are properly copied or shared.

Originally, we used a shared pointer to ensure that the stream info
used in deferred logging could outlive the stream itself. However,
at least one member of stream info is a raw pointer (request headers)
and runs the risk of pointing at a destroyed object. In the current
implementation, this would have been safe, but having stream info
be a shared pointer opens us up to similar problems in the future.
The explicit copy is a safeguard against that situation.

Signed-off-by: Paul Sohn <paulsohn@google.com>