util: make OptRef not derive from absl::optional<std::reference<>> to reduce size footprint from 16 bytes to 8. (#24661)
Commit Message: I wrote OptRef to add syntactic sugar to users of optional references, which were very awkward to program with. However @pianiststickman recently observed with sadness that it consumes aligned(8+1) bytes, ie 16 bytes, where in previous times we would just used a pointer and gone on with our lives :) This often doesn't matter unless we find OptRefs in structs with high cardinality.

This change drops the inheritance from optional<reference_wrapper<>> and just implements the required functionality directly. Some of the functionality inherited from absl::optional was hard to replicate, and was generally getting in the way of correctness, performance, or readability. So in a few cases (7 files), small modifications to call-sites had to be supplied.

Notable among them:

It was possible to assign optional<reference_wrapper<T>> to an optional<T> which would copy the object if it has a value. I changed that to require calling a copy() method, as I think the previous value-copy was non-obvious and should be explicit.
There is some property of optional<reference_wrapper<T>> where you could take a const optional<reference_wrapper<T>> and extract out a non-const T&. I think this does not make sense and didn't want to replicate it in this container object. This was exploited in a few places. In some of them it was easy to change the call-sites to use a const T& but in one of them I added an explicit const_cast at the call-site where this was needed.
This change duplicates some of the methods from absl::optional that were previously inherited, and thus explicit tests were added to ensure it the wrapper is covered completely with its own unit test, which I verified manually in the debugger.

Additional Description:
Risk Level: low -- this is a low level class but it is well tested and generally errors should be found by the compiler.
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

Signed-off-by: Joshua Marantz <jmarantz@google.com>
Signed-off-by: VishalDamgude <vishal.damgude@freshworks.com>