Make collection of opentelemetry traces optional
Mark the `grpc_service` of the opentelemtry configuration as optional
and if the resulting field is empty, abstain from sending the trace data
to any collection service.

This means that the opentelemetry plugin will still generate and
propagate trace headers, but they will no longer be sent to the
collector.

Signed-off-by: Ashish Banerjee <ashish.banerjee@solo.io>