bazel: fix missing framework on macOS
When using only the command line tools (not a full Xcode installation)
on macOS, bazel does not pass `-fobjc-link-runtime` and therefore does
not link to the CoreFoundation framework, which is required by absl.
This adds the required framework, and makes this a linker error instead
of a runtime error for the future.

Fixes https://github.com/envoyproxy/envoy/issues/24782

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>