Fix tcp_proxy: mark done reading on upstream trailers (#24800)
When using HTTP/2 upstream with TCP tunneling feature, the only place the reading is done (i.e using "doneReading()" callback) is in cases the last data frame from the upstream also contains the NGHTTP2_FLAG_END_STREAM flag. This way the decodeData callback will have "end_stream" to true.

In certain streaming cases, the downstream sends delayed FIN packet and the server will not respond with a data frame, rather than a headers frame which will be the trailers data and an END_STREAM flag. This causes "doneReading()" callback to be skipped and the encoder would not be reset, which causes the filter chain not to drain.

Fixes #24704

Signed-off-by: ohadvano <49730675+ohadvano@users.noreply.github.com>