mobile: register all Envoy Mobile extension factories explicitly
In preparation for disabling static extension factory registration in
https://github.com/envoyproxy/envoy/pull/24844, start by explicitly
force-registering the extension factories Envoy Mobile is currently
using.

To do this, some of the extension factories that were previously only
registered also need to be declared with `DECLARE_FACTORY` so its
`forceRegister...` function can be made accessible externally.

This is effectively a no-op since all these extension factories should
already be registered statically.

Signed-off-by: JP Simard <jp@jpsim.com>