extensions: split extension registration macro
In preparation for disabling static extension factory registration in
https://github.com/envoyproxy/envoy/pull/24844, split the
`REGISTER_FACTORY` macro into two, one that doesn't take a deprecated
name and one that does, called `REGISTER_FACTORY_D` where the `_D`
indicates that it accepts a deprecated name.

This is done because the previous approach of appending a
`{deprecated_name}` suffix at the usage site is incompatible with the
approach taken by https://github.com/envoyproxy/envoy/pull/24844 to not
statically register the extension factories.

Signed-off-by: JP Simard <jp@jpsim.com>