mobile: reducing the number of enabled Envoy stats (#24980)
Commit Message: As a result of changes from https://github.com/envoyproxy/envoy/pull/24733 we ended up increasing the number of enabled stats from 317 to 462. We did not really want to increase the number of enabled stats but doing this was required if we wanted our stats inclusion list to use as many `prefix`es that end up with a dot "." character as possible. That kind of `prefix`es are optimized in a special way that we hoped would help us to improve envoy mobile launch times. Since then, we learnt that as they are now upload of Envoy Mobile stats consumes a lot of bandwidth and we are working on balancing the app launch time with the use of network bandwidth - for this reason, we decided to disable some of the stats enabled in https://github.com/envoyproxy/envoy/pull/24733. We are not going back to using regexes - instead we use prefixes.
Additional Description:
Risk Level: Low, less of enabled stats at a cost of slightly longer Envoy launch time.
Testing:
Docs Changes:
Release Notes:
Platform Specific Features:
[Optional Runtime guard:]
[Optional Fixes #Issue]
[Optional Fixes commit #PR or SHA]
[Optional Deprecated:]
[Optional [API Considerations](https://github.com/envoyproxy/envoy/blob/main/api/review_checklist.md):]