xds: Make AGGREGATED_DELTA_GRPC set is_aggregated on the Subscription
Currently, there's a bug when the ApiConfigSource is
AGGREGATED_DELTA_GRPC and xdstp is used: The
GrpcCollectionSubscriptionImpl that gets created sets is_aggregated to
false, instead of setting it to true.  This causes the
GrpcSubscriptionImpl instance to attempt to start the GrpcMux on start,
instead of waiting to share a gRPC stream like ADS is supposed to.

Unfortunately, this behavior was masked by the integration tests,
because we use a FakeUpstream in the integration tests that just
contains a single xDS stream to which the Envoy instances connect to.

The only way to verify this behavior was to look at the logs when
running the XdsTpAdsIntegrationTest: when is_aggregated was set to false
(prior to this commit), the logs would contain GrpcStream messages
saying the stream already exists.  After changing is_aggregated to true,
the logs no longer output stream "already exists", because we don't
attempt to create multiple streams, which is the correct behavior.

Signed-off-by: Ali Beyad <abeyad@google.com>