Move the old-android implementation of getifaddrs from Envoy Mobile to Envoy code (#24983)
Move the old-android implementation of getifaddrs from Envoy Mobile to Envoy code
Before this PR, Envoy had an implementation of getifaddrs which only worked on relatively recent version of android (24, 25, 26, etc). It also had an API for an embedder to supply an alternate implementation which would be used instead of the core implementation. Meanwhile, Envoy Mobile had and implementation of getifaddrs which worked on older version of android (23, 22, 21, etc). This code moves that Envoy Mobile "alternate" implementation into core Envoy leaving Envoy with support for any version of Android. Also, since Envoy Mobile was the only user of Envoy's "alternate getifaddrs implementation" codepath, this PR also removes the now unused setAlternateGetifaddrs() method.

This is an alternative to #24907

Risk Level: Low
Testing: Unit
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: Provides a getifaddrs implementation for old versions of Android.

Signed-off-by: Ryan Hamilton <rch@google.com>