filter state: share once with upstream (#24920)
Commit Message: Add an option to avoid transitive sharing of the filter state objects with the upstream filter state.
Additional Description: Consider a chain of two internal listeners: decap (socket listener), HTTP listener (internal), encap (internal). A shared filter state from the decap listener is propagated to both the HTTP and the encap listener. That's not always desirable if the filter state is processed in the HTTP listener because it forces additional pools for the connections between the HTTP and encap listener. This option allows copying once from the decap to the HTTP listener without affecting the pooling between the HTTP and encap listeners.
Risk Level: low
Testing: unit
Docs Changes: none
Release Notes: none