Use route name suffix to specify header only decoding.
Context: we have a requirement to support http backends. But we still
need to wave the information in the path to proto message in order to
develop the features on top of proto message, for example, resource
extraction for IAM authz; field scrubbing for audit logging.

Therefore we came up with this idea to perform "header-only" decoding so
we can offer the above features without touching bodies or any response
data.

We have a subsequent filter which will read the decoded header in the
filter state and will process accordingly.

This is an early draft to seek feedback. Let us know how you think about
it. Thank you.