quic: deferred logging of roundtrip response time (#23648)
Commit Message: For QUIC, defer access logging to when the final ack is received from downstream.
Additional Description:
This PR implements QuicAckListenerInterface which allows QUIC streams to listen for acks. Here, we use the ack listener to record a "roundtrip response time" that is analogous to the full response time or time-to-last-byte as experienced by a downstream client. We also defer access logging to the ack listener in order to record this metric in the stream info and make it available to access logs. The stream info is copied into the ack listener so that it can be used for logging even after its originating stream is destroyed.

Risk Level: Medium
Testing: Integration tests.
Docs Changes: N/A
Release Notes: added
Platform Specific Features: N/A
Runtime guard: envoy_reloadable_features_quic_defer_logging_to_ack_listener defaults to true.

Signed-off-by: Paul Sohn <paulsohn@google.com>