Increase header size limit in DownstreamProtocolIntegrationTest.ManyTrailerHeaders.
This is because BalsaParser includes every byte of headers (and
trailers) in the size that it compares to the limit.  This includes ": "
separator between key and header, and "\r\n" after each header line.  In
contrast, http-parser does not have built-in size limit enforcement, and
limit is incremented in ConnectionImpl using HeaderMapImpl::byteSize(),
which only considers key and value lengths.

Signed-off-by: Bence Béky <bnc@google.com>