[file_system_http_cache] Add eviction (#24642)
Commit Message: [file_system_http_cache] Add eviction
Additional Description: Add the last piece of making file_system_http_cache functional; eviction. This adds a thread for the eviction operation to be performed on, allowing for on-demand eviction rather than periodic. This has advantages and disadvantages

Pro: when there is no demand for eviction because the cache is below its size limits, the cost is very low.
Pro: the cache should not get significantly oversized, because a demand for eviction will occur as soon as it is oversized.
Con: if the cache is constantly being populated and entries rotated out, the cost will be higher than periodic eviction would be.
Option for improvement: a configuration option for postponing eviction if the last eviction was more recent than (specified duration) would reduce the cost of that con, but reduce the benefit of the second pro.
Risk Level: None, this filter is not in use and is still tagged WIP.
Testing: Almost 100% coverage (updating an atomic more than once because there was a race is not covered).
Docs Changes: Adds DESIGN.md to explain the various eviction strategy decisions and other implementation details.
Release Notes: n/a (not yet)
Platform Specific Features: n/a

Signed-off-by: Raven Black <ravenblack@dropbox.com>