mobile: add support for direct responses to C++ builder
And start validating yaml comparison for Swift direct response tests.

Skip C++ assertions for builder configurations that use direct responses
since it's going to be a lot of work to generate that YAML, it's
only ever used for tests and it's meant as a temporary transition
helper.

This helps migrate the ObjC/Swift APIs off the custom yaml codepaths.

Signed-off-by: JP Simard <jp@jpsim.com>