mobile: change request compression to be per-request
In practice it's very likely for mobile apps to make HTTP requests to
endpoints they don't control and that may not support request
decompression at all, or may not support certain compression schemes
like Brotli.

So this change reworks Envoy Mobile's request compression configuration
to operate on a per-request basis by adding a new
`RequestHeadersBuilder.enableRequestCompression(...)` function where
either gzip or Brotli can be specified as the compression algorithm. Not
calling this function does not compress the request at all.

The `enableGzipCompression` and `enableBrotliCompression` functions on
engine builders have been removed.

The `enableRequestCompression` function is only available when building
with `--define=envoy_mobile_request_compression=enabled`, which is the
default.

Signed-off-by: JP Simard <jp@jpsim.com>