[balsa] Use ProcessHeaders() instead of OnHeader(). (#25475)
BalsaVisitorInterface provides two ways of processing headers:
OnHeader() is called for each header line, and ProcessHeaders() and
ProcessTrailers() are called with the entire header/trailer block.  All
of these methods are only called after the entire header block or
trailer block is parsed, so using OnHeader() does not actually allow
early processing of headers.

The major difference is that some header validation happens after
OnHeader() is called, for example, check for empty header names (both
for headers and trailers), multiple Content-Length headers, multiple
Transfer-Encoding headers, and the presence of Content-Length and
Transfer-Encoding in the same header block (not for trailers).  (Some of
these test can be disabling using HttpValidationPolicy.)  If the headers
fail, for example, because of an empty field name or because of multiple
Transfer-Encoding headers, ProcessHeaders() or ProcessTrailers() are not
called.

One problem with using OnHeader() is that an assertion in
ConnectionImpl::completeCurrentHeader() can fail if the header name is
empty, see issue #25458.  Another problem is that header names and
values are copied immediately, using unnecessary memory in case an error
is detected later.  This PR migrates BalsaParser from using OnHeader()
to ProcessHeaders() and ProcessTrailers().

This PR also adds two tests.  The first one, the case of empty header
name, would cause an assertion failure without the change.  The second
one would pass either way, because the error from Balsa is propagated
properly to ConnectionImpl.  Note that the error message is slightly
different with http-parser, because http-parser does not have a check
for Transfer-Encoding, but ConnectionImpl::onHeadersCompleteImpl()
signals an error, because the combined value of "chunked,chunked" does
not match "chunked".

Relevant issues: #25458, #21245.

Signed-off-by: Bence Béky <bnc@google.com>