[balsa] Enable custom request methods.
This long-standing feature request is one of the major drivers of the
http-parser to Balsa migration.  It makes sense to unconditionally
enable custom methods (unless we have evidence that this causes
problems, in which case a separate knob can be added).

Also remove a fistful of unnecessary absl::StrCat() calls from
codec_impl_test.cc.

Signed-off-by: Bence Béky <bnc@google.com>