filter-state: trigger ENVOY_BUG instead of throw to avoid crashing
Previously, an API error around filter-state (by double-setting a
read-only key, or setting data with different scopes) would result in
an exception being thrown, which would most commonly result in Envoy
crashing.

Switch this to ENVOY_BUG, so that the API mis-use can be noticed, but
without crashing the process.

This more closely aligns with Envoy's policy of not using exceptions
in the data path: filter state is set almost exclusively in the
request process path.

Signed-off-by: Greg Greenway <ggreenway@apple.com>