test: Move expect assertions to prevent flakes
Move expect assertions after the first waitForData() to be sure
onNewConnection() callback has had chance to schedule. This solves the
~10% flakiness caused by expecting onNweConnection() haveing been called
before the first waitForData(). Note that at the point of the expectation
read callback (onData()) has not been called on the connection yet, so
onNewConnection() has not been called if the runtime guard
`envoy.reloadable_features.initialize_upstream_filters` is set to
`false`.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>