tls: add option to set signature algorithms
Add a new optional field to TlsParameters to configure signature algorithms.
The value is used during TLS handshake in TLS Signature Algorithms extension.
It indicates which signature algorithms may be used for digital signatures.
The value can be used for both DownstreamTlsContext and UpstreamTlsContext.

When user does not set the option, the default value is defined by Envoy.
The current default value is selected according to what BoringSSL uses
internally as default.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>