listener filters: improve logging levels
Currently, we have a generic "new connection accepted" at debug level,
but "set to {}" at trace. This ends up being the inversion of what we
typically want; we know any new connection will call the listener
filter, but what is interesting is knowing the result.

This PR simply inverts the scopes, so we have the same amount of logs at
`debug` but get more information.

Before this PR, a typical log would look like:

```
{"level":"debug","msg":"original_dst: new connection accepted"}
{"level":"trace","msg":"original_dst: set destination to 1.2.3.4:80"}
```

Inverting this will have a big impact for us, as folks usually send us `debug` level logs,
which skips this critical bit of info

Signed-off-by: John Howard <howardjohn@google.com>