[mobile] ci: switch to `pull_request_target`
This changes the PR trigger from `pull_request` to `pull_request_target`.
The latter is executed in the context of the PR target (usually `main`) and has access to secrets stored in GitHub, while
the former is executed in the context of the PR head, which is potentially in a fork and thus doesn
get access to secrets.

This is part of the effort to reduce flakyness in the CI build because of authentication problems against
the EngFlow Remote Execution cluster.

Signed-off-by: Yannic Bonenberger <contact@yannic-bonenberger.com>