mobile: add CI job to publish aars to Sonatype
Schedule it so it runs automatically on Mondays at 1pm UTC (8am EST).

This pulls from the CI jobs that were set up in the previous Envoy
Mobile repo:

* https://github.com/envoyproxy/envoy-mobile/blob/main/.github/workflows/release.yml
* https://github.com/envoyproxy/envoy-mobile/blob/main/.github/workflows/weekly_release.yml

However, we don't want to create git tags or GitHub releases so not to
interfere with the main Envoy tags and releases.

Signed-off-by: JP Simard <jp@jpsim.com>