Reduce Cluster memory utilization by avoiding Metadata instances when not needed (#26231)
Reduce Cluster memory utilization by avoiding Metadata instances when not needed

Currently an instance of raw Metadata and an instance of TypedMetadata are created in each cluster, even if no metadata is configured. This PR switches instances to unique_ptr and only creates Metadata objects when actually configured. Otherwise a singleton default metadata is used.

This is a step towards more memory efficient config data structures. Issue #24154.

A similar change was made to the Route in #24309. This PR is consistent with the approach applied there.

Signed-off-by: Yury Kats <ykats@google.com>