Use PackedStruct to store optional timeouts in Route
In #24659 optional Route timeouts were optimized by using a struct
instead of `absl::optional`. Since then, a more efficient `PackedStruct`
was introduced, that provides even more efficiency, especially in the
cases when some timeouts are set, but not all. The PackedStruct is  storing only set values, and not waisting 8 bytes on each of them being optional.

This PR is switching Routes from using a internal struct to using
PackedStruct.

This is a step towards more memory efficient config data structures. Issue #24154.

Signed-off-by: Yury Kats <ykats@google.com>