datadog: Tracer portion of new tracing library (#26148)
This is part of the work proposed in #23958.

This adds the Tracer class, which has one member function, startSpan, for creating a new trace or extracting one from incoming (HTTP) request headers.

Tracer is a thread-local wrapper around the corresponding class in dd-trace-cpp, except that failure to configure the underlying tracer puts Tracer into a "no-op mode" that always produces NullSpan instances.

In addition to that code and its corresponding unit tests, I also updated the version of dd-trace-cpp used. The newer version has additional Span methods that made testing easier. Additionally, the newer dd-trace-cpp models TraceID as its own class to accommodate 128-bit IDs. The existing unit test for Span needed to be modified to account for this.

Signed-off-by: David Goffredo <david.goffredo@datadoghq.com>
Signed-off-by: River Phillips <riverphillips1@gmail.com>