Golang: Improve error handling
While working on the golang filter I saw the following improvements we can make
in the error handling:

1. Expose stats for panics and generally adding `stats.h` for the golang filter
2. Surface panic messages to envoy error logs

This is achieved by moving the implementation of panic error handling from `SendLocalReply` to
a dedicated function that can add the o11y for the errors before we surface them to the users.

Also removed `Golang` from the response message when the golang filter panics to avoid leaking implementation
details to end users.

Signed-off-by: Sotiris Nanopoulos <sotirisnan@gmail.com>